Expose `save` function from records-queue
This replaces the dynamic function call in controller-webform, making it a little easier to find the call site. If at some point in the future we migrate to TypeScript this would be somewhat less necessary.

Another option, discussed with @lognaturel, would be to remove this logic from controller-webform and have records-queue determine which is appropriate. I agree this would be an improvement, but I think that's probably out of scope for the last-saved work.