Support for jr:instance/last-saved endpoint (#269)
* Pin Node version with Volta

* Initial form-cache implementation of get/set last-saved

* Don't set last saved records when encrypted

* Expose `save` function from records-queue

This replaces the dynamic function call in controller-webform, making it a little easier to find the call site. If at some point in the future we migrate to TypeScript this would be somewhat less necessary.

Another option, discussed with @lognaturel, would be to remove this logic from controller-webform and have records-queue determine which is appropriate. I agree this would be an improvement, but I think that's probably out of scope for the last-saved work.

* Implement/test offline support for last-saved…

This includes a refactor of the encryption exception, ensuring Surveys with encryption enabled do not store unencrypted draft Records. This refactor was accomplished by persisting encryption-enabled state on the Survey itself.

* Support online mode, populate last-saved secondary instances

* Correct typo

* Initial changes for feedback on PR #269

- Corrected assignment to `binaryDefaults`
- Corrected a typo
- Removed references to last-saved where no longer relevant

* Move tests related to last-saved to test/client/feature/last-saved.spec.js

* Move tests related to encryption-enabled surveys/last-saved records to test/client/feature/survey-encryption.spec.js

* Document fallback to a survey's model when it has no last-saved record

* Add AbortController polyfill for old Chrome version in CI

* Don't store last saved records for surveys without a last-saved secondary instance

* Polyfill Object.fromEntries...

* Blob.text -> Response.text to hopefully address CI failure

* Fix overwrite of lastSavedRecord on first online submit, edit regression

* Move remaining last-saved tests to the feature

* Fix `npm run build-docs`

* Use, and fix, getDisplayableRecordList