Add VSCode test/launch debugging task
This must be performed in two tasks because Karma is a Node process which launches a browser child process, which the VSCode debugger cannot attach to directly. So debugging tests:

1. Launches the `test-watch` process
2. Waits until the `karma:watch` subprocess begins
3. Backgrounds the `test-watch` process
4. Attaches the debugger to the Chrome process launched by Karma

Caveat: there is some delay in step 4. In some startup cases where a breakpoint is encountered/evaluated before the debugger attaches, a rerun will be necessary to hit that breakpoint.