Fix: restore population of deprecatedID in edit mode
This was a regression in #321. `record` here is populated by `_checkAutoSavedRecord`, which is only available in offline mode. Since that will always be falsey (undefined) in edit mode, `submitted` would be set to false. Since we should only set `deprecatedID` when editing submissions, it made more sense to assign it directly based on the `isEditing` property introduced in #269 (which should also have been added to the JSDoc type at that time).